Updating layers to produce better Galleon integration and some default configuration.
* Updating the layers
* Moving the in memory embedding store from a file to a path
* Splitting the in memory embedding models to be able to trim more effectively
* Default embedding content retriever
* Renaming all the resources
