--- ai-feature-pack/src/main/resources/layers/standalone/embeddings/layer-spec.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright The WildFly Authors
-  ~ SPDX-License-Identifier: Apache-2.0
-  -->
-<layer-spec xmlns="urn:jboss:galleon:layer-spec:2.0" name="embeddings">
-    <dependencies>
-        <layer name="ai"/>
-    </dependencies>
-    <packages>
-        <package name="dev.langchain4j.embeddings.all-minilm-l6-v2"/>
-        <package name="dev.langchain4j.embeddings.all-minilm-l6-v2-q"/>
-        <package name="dev.langchain4j.embeddings.bge-small-en"/>
-        <package name="dev.langchain4j.embeddings.bge-small-en-q"/>
-        <package name="dev.langchain4j.embeddings.bge-small-en-v15"/>
-        <package name="dev.langchain4j.embeddings.bge-small-en-v15-q"/>
-        <package name="dev.langchain4j.embeddings.e5-small-v2"/>
-        <package name="dev.langchain4j.embeddings.e5-small-v2-q"/>
-    </packages>
-</layer-spec>
