+++ ai-feature-pack/src/main/resources/layers/standalone/in-memory-embedding-model-bge-small-en/layer-spec.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+~ Copyright The WildFly Authors
+~ SPDX-License-Identifier: Apache-2.0
+-->
+<layer-spec xmlns="urn:jboss:galleon:layer-spec:2.0" name="in-memory-embedding-model-bge-small-en">
+    <props>
+        <prop name="org.wildfly.rule.add-on-depends-on" value="only:ai"/>
+        <prop name="org.wildfly.rule.add-on" value="ai-embedding-model,in-memory-embedding-model-bge-small-en"/>
+        <prop name="org.wildfly.rule.add-on-description" value="BAAI bge-small-en embedding model that runs within your Java application's process."/>
+    </props>
+    <dependencies>
+        <layer name="ai"/>
+    </dependencies>
+    <packages>
+        <package name="dev.langchain4j.embeddings.bge-small-en"/>
+    </packages>
+    <feature spec="subsystem.ai">
+        <feature spec="subsystem.ai.in-memory-embedding-model">
+            <param name="in-memory-embedding-model" value="bge-small-en"/>
+            <param name="module" value="dev.langchain4j.embeddings.bge-small-en"/>
+            <param name="embedding-class" value="dev.langchain4j.model.embedding.onnx.bgesmallen.BgeSmallEnEmbeddingModel"/>
+        </feature>
+    </feature>
+</layer-spec>
