+++ ai-feature-pack/src/main/resources/layers/standalone/in-memory-embedding-model-e5-small-v2/layer-spec.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+~ Copyright The WildFly Authors
+~ SPDX-License-Identifier: Apache-2.0
+-->
+<layer-spec xmlns="urn:jboss:galleon:layer-spec:2.0" name="in-memory-embedding-model-e5-small-v2">
+    <props>
+        <prop name="org.wildfly.rule.add-on-depends-on" value="only:ai"/>
+        <prop name="org.wildfly.rule.add-on" value="ai-embedding-model,in-memory-embedding-model-e5-small-v2"/>
+        <prop name="org.wildfly.rule.add-on-description" value="Microsoft E5-small-v2 embedding model that runs within your Java application's process."/>
+    </props>
+    <dependencies>
+        <layer name="ai"/>
+    </dependencies>
+    <packages>
+        <package name="dev.langchain4j.embeddings.e5-small-v2"/>
+    </packages>
+    <feature spec="subsystem.ai">
+        <feature spec="subsystem.ai.in-memory-embedding-model">
+            <param name="in-memory-embedding-model" value="e5-small-v2"/>
+            <param name="module" value="dev.langchain4j.embeddings.e5-small-v2"/>
+            <param name="embedding-class" value="dev.langchain4j.model.embedding.onnx.e5smallv2.E5SmallV2EmbeddingModel"/>
+        </feature>
+    </feature>
+</layer-spec>
