--- ai-feature-pack/src/main/resources/layers/standalone/ollama/layer-spec.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright The WildFly Authors
-  ~ SPDX-License-Identifier: Apache-2.0
-  -->
-<layer-spec xmlns="urn:jboss:galleon:layer-spec:2.0" name="ollama">
-    <dependencies>
-        <layer name="ai"/>
-    </dependencies>
-    <packages>
-        <package name="dev.langchain4j.ollama"/>
-    </packages>
-</layer-spec>
