+++ ai-feature-pack/src/main/resources/layers/standalone/openai-chat-model/layer-spec.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+~ Copyright The WildFly Authors
+~ SPDX-License-Identifier: Apache-2.0
+-->
+<layer-spec xmlns="urn:jboss:galleon:layer-spec:2.0" name="openai-chat-model">
+    <props>
+        <prop name="org.wildfly.rule.class" value="dev.langchain4j.model.chat.ChatLanguageModel"/>
+        <prop name="org.wildfly.rule.add-on-depends-on" value="only:ai"/>
+        <prop name="org.wildfly.rule.add-on" value="ai-llm,openai-chat-model"/>
+        <prop name="org.wildfly.rule.add-on-description" value="Use OpenAI API for LLM interactions"/>
+    </props>
+    <dependencies>
+        <layer name="ai"/>
+    </dependencies>
+    <packages>
+        <package name="dev.langchain4j.openai"/>
+    </packages>
+    <feature spec="subsystem.ai">
+        <feature spec="subsystem.ai.openai-chat-model">
+            <param name="openai-chat-model" value="openai"/>
+            <param name="base-url" value="${org.wildfly.ai.openai.chat.url,env.OPENAI_CHAT_URL:https://api.groq.com/openai/v1}"/>
+            <param name="api-key" value="${env.GROQ_API_KEY}"/>
+            <param name="model-name" value="${org.wildfly.ai.openai.chat.model.name,env.OPENAI_CHAT_MODEL_NAME:llama3-8b-8192}"/>
+            <param name="log-requests" value="${org.wildfly.ai.openai.chat.log.request,env.OPENAI_CHAT_LOG_REQUEST:true}"/>
+            <param name="log-responses" value="${org.wildfly.ai.openai.chat.log.response,env.OPENAI_CHAT_LOG_RESPONSE:true}"/>
+        </feature>
+    </feature>
+</layer-spec>
