--- ai-feature-pack/src/main/resources/layers/standalone/weaviate/layer-spec.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright The WildFly Authors
-  ~ SPDX-License-Identifier: Apache-2.0
-  -->
-<layer-spec xmlns="urn:jboss:galleon:layer-spec:2.0" name="weaviate">
-    <dependencies>
-        <layer name="ai"/>
-    </dependencies>
-    <packages>
-        <package name="dev.langchain4j.weaviate"/>
-    </packages>
-</layer-spec>
