+++ wildfly-ai/subsystem/src/main/java/org/wildfly/extension/ai/AISubsystemRegistrar.java
@@ -12,16 +12,16 @@ import org.jboss.as.controller.descriptions.ParentResourceDescriptionResolver;
 import org.jboss.as.controller.descriptions.SubsystemResourceDescriptionResolver;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
 import org.jboss.as.server.deployment.Phase;
-import org.wildfly.extension.ai.model.chat.OllamaChatLanguageModelProviderRegistrar;
-import org.wildfly.extension.ai.model.chat.OpenAIChatLanguageModelProviderRegistrar;
+import org.wildfly.extension.ai.chat.OllamaChatLanguageModelProviderRegistrar;
+import org.wildfly.extension.ai.chat.OpenAIChatLanguageModelProviderRegistrar;
 import org.wildfly.extension.ai.deployment.AIDependencyProcessor;
 import org.wildfly.extension.ai.deployment.AIDeploymentProcessor;
-import org.wildfly.extension.ai.model.embedding.EmbeddingModelProviderRegistrar;
-import org.wildfly.extension.ai.store.embedding.InMemoryEmbeddingStoreProviderRegistrar;
-import org.wildfly.extension.ai.model.embedding.OllamaEmbeddingModelProviderRegistrar;
+import org.wildfly.extension.ai.embedding.model.InMemoryEmbeddingModelProviderRegistrar;
+import org.wildfly.extension.ai.embedding.store.InMemoryEmbeddingStoreProviderRegistrar;
+import org.wildfly.extension.ai.embedding.model.OllamaEmbeddingModelProviderRegistrar;
 import org.wildfly.extension.ai.rag.retriever.EmbeddingStoreContentRetrieverProviderRegistrar;
 import org.wildfly.extension.ai.rag.retriever.WebSearchContentContentRetrieverProviderRegistrar;
-import org.wildfly.extension.ai.store.embedding.WeaviateEmbeddingStoreProviderRegistrar;
+import org.wildfly.extension.ai.embedding.store.WeaviateEmbeddingStoreProviderRegistrar;
 import org.wildfly.subsystem.resource.ManagementResourceRegistrar;
 import org.wildfly.subsystem.resource.ManagementResourceRegistrationContext;
 import org.wildfly.subsystem.resource.ResourceDescriptor;
@@ -50,7 +50,7 @@ class AISubsystemRegistrar implements SubsystemResourceDefinitionRegistrar {
         ManagementResourceRegistrar.of(descriptor).register(registration);
         new OpenAIChatLanguageModelProviderRegistrar(RESOLVER).register(registration, context);
         new OllamaChatLanguageModelProviderRegistrar(RESOLVER).register(registration, context);
-        new EmbeddingModelProviderRegistrar(RESOLVER).register(registration, context);
+        new InMemoryEmbeddingModelProviderRegistrar(RESOLVER).register(registration, context);
         new OllamaEmbeddingModelProviderRegistrar(RESOLVER).register(registration, context);
         new InMemoryEmbeddingStoreProviderRegistrar(RESOLVER).register(registration, context);
         new WeaviateEmbeddingStoreProviderRegistrar(RESOLVER).register(registration, context);
