+++ wildfly-ai/subsystem/src/main/resources/org/wildfly/extension/ai/LocalDescriptions.properties
@@ -16,16 +16,17 @@ ai.embedding-store-content-retriever.min-score=The minimum relevance score for t
 ai.embedding-store-content-retriever.max-results=The maximum number of contents to retrieve.
 ai.embedding-store-content-retriever.remove=This operation removes an embedding store content retriever.
 
-ai.embedding-model=Embedding model for AI
-ai.embedding-model.add=This operation adds an embedding model
-ai.embedding-model.remove=This operation removes an embedding model
-ai.embedding-model.module=The module to load the embedding model from
-ai.embedding-model.embedding-class=The class from which the embedding model will be instanciated
+ai.in-memory-embedding-model=Embedding model for AI
+ai.in-memory-embedding-model.add=This operation adds an embedding model
+ai.in-memory-embedding-model.remove=This operation removes an embedding model
+ai.in-memory-embedding-model.module=The module to load the embedding model from
+ai.in-memory-embedding-model.embedding-class=The class from which the embedding model will be instanciated
 
 ai.in-memory-embedding-store=In memory embedding store for AI.
 ai.in-memory-embedding-store.add=This operation adds an in memory embedding store.
 ai.in-memory-embedding-store.remove=This operation removes an in memory embedding store.
-ai.in-memory-embedding-store.file=The file to load the in memory embedding store content from.
+ai.in-memory-embedding-store.path=The actual filesystem path to be loaded as content of the in-memory embedding store. Treated as an absolute path, unless the 'relative-to' attribute is specified, in which case the value is treated as relative to that path.
+ai.in-memory-embedding-store.relative-to=Reference to a filesystem path defined in the "paths" section of the server configuration.
 
 ai.ollama-chat-model=Chat model for AI for Ollama.
 ai.ollama-chat-model.add=This operation adds a chat model
